Refactor : RoomHeader 컴포넌트 분리 - Hamburger, Code, Avatar, Door로 분리 Issue #525
